refactor: reorganizar estrutura de arquivos para pasta /app
- Mover páginas de remessas e tanques para /app/app/dashboard/
- Atualizar rotas no sidebar para novos caminhos
- Remover arquivos antigos da estrutura /dashboard
